Split dependencies into multiple workspaces and rename .storybook to storybook folder.
