+++ .github/workflows/storybook.yml
@@ -13,8 +13,8 @@ on:
       - '!assets/**/__tests__/**/*.js'
       - '!assets/**/test/*.js'
       - '!assets/**/*.test.js'
-      - '.storybook/**'
-      - '!.storybook/storybook-data.js'
+      - 'storybook/**'
+      - '!storybook/storybook-data.js'
       - '**.scss'
       - '.nvmrc'
       - '**/package.json'
@@ -32,8 +32,8 @@ on:
       - '!assets/**/__tests__/**/*.js'
       - '!assets/**/test/*.js'
       - '!assets/**/*.test.js'
-      - '.storybook/**'
-      - '!.storybook/storybook-data.js'
+      - 'storybook/**'
+      - '!storybook/storybook-data.js'
       - '**.scss'
       - '.nvmrc'
       - '**/package.json'
