+++ assets/package.json
@@ -1,0 +1,70 @@
+{
+	"dependencies": {
+		"@jackfranklin/test-data-bot": "^2.1.0",
+		"@lit-labs/react": "^1.2.1",
+		"@material-ui/core": "^4.12.4",
+		"@material/button": "^2.3.0",
+		"@material/checkbox": "^2.3.0",
+		"@material/dialog": "^2.3.0",
+		"@material/form-field": "^2.3.0",
+		"@material/layout-grid": "^0.41.0",
+		"@material/linear-progress": "^1.1.0",
+		"@material/list": "^2.3.0",
+		"@material/menu": "^2.3.0",
+		"@material/radio": "^2.3.0",
+		"@material/react-card": "^0.15.0",
+		"@material/react-chips": "^0.15.0",
+		"@material/react-dialog": "^0.15.0",
+		"@material/react-select": "^0.15.0",
+		"@material/react-tab-bar": "^0.15.0",
+		"@material/react-text-field": "^0.15.0",
+		"@material/ripple": "^2.3.0",
+		"@material/switch": "^2.3.0",
+		"@material/web": "^1.5.1",
+		"@reach/combobox": "^0.18.0",
+		"@react-hook/merged-ref": "^1.3.2",
+		"@react-hook/window-size": "^3.1.1",
+		"@shopify/jest-dom-mocks": "^5.2.0",
+		"@storybook/addon-queryparams": "^6.2.9",
+		"@testing-library/react": "^10.4.9",
+		"@wordpress/api-fetch": "^3.23.1",
+		"@wordpress/compose": "^3.25.3",
+		"@wordpress/data": "^4.27.3",
+		"@wordpress/dom-ready": "^2.13.2",
+		"@wordpress/element": "^2.20.3",
+		"@wordpress/i18n": "^3.20.0",
+		"@wordpress/icons": "^2.10.3",
+		"@wordpress/keycodes": "^3.58.0",
+		"@wordpress/url": "^3.59.0",
+		"classnames": "^2.5.1",
+		"clipboard-copy": "^3.2.0",
+		"compare-versions": "^3.6.0",
+		"dompurify": "^2.5.6",
+		"faker": "^5.5.3",
+		"fetch-mock": "^9.11.0",
+		"focus-trap-react": "^10.2.3",
+		"google-pax-sdk": "^1.0.12",
+		"immer": "^9.0.21",
+		"intersection-observer": "^0.12.2",
+		"invariant": "^2.2.4",
+		"jsdom": "^16.7.0",
+		"just-detect-adblock": "^1.1.0",
+		"lodash": "^4.17.21",
+		"md5": "^2.3.0",
+		"memize": "^1.1.0",
+		"patch-package": "^8.0.0",
+		"prop-types": "^15.8.1",
+		"psl": "^1.9.0",
+		"punycode": "^2.3.1",
+		"query-string": "^7.1.3",
+		"react": "^16.14.0",
+		"react-dom": "^16.14.0",
+		"react-google-charts": "^3.0.15",
+		"react-joyride": "^2.8.2",
+		"react-router-dom": "^5.3.4",
+		"react-use": "^15.3.8",
+		"react-use-observer": "^2.2.4",
+		"rxjs": "^6.6.7",
+		"use-memo-one": "^1.1.3"
+	}
+}
