+++ stories/package.json
@@ -1,0 +1,5 @@
+{
+	"dependencies": {
+		"@storybook/react": "^6.5.16"
+	}
+}
