+++ tests/backstop/package.json
@@ -1,0 +1,8 @@
+{
+	"dependencies": {
+		"@componentdriven/csf": "0.0.2-alpha.0",
+		"glob": "^7.2.3",
+		"lodash": "^4.17.21",
+		"node-fetch": "^2.7.0"
+	}
+}
