+++ tests/backstop/scenarios.js
@@ -34,11 +34,11 @@ const path = require( 'path' );
 /**
  * Internal dependencies
  */
-const legacyStorybookScenarios = require( '../../.storybook/storybook-data' );
-const storybookConfig = require( '../../.storybook/main' );
+const legacyStorybookScenarios = require( '../../storybook/storybook-data' );
+const storybookConfig = require( '../../storybook/main' );
 const rootURL = 'file:///src/dist/iframe.html?id=';
 
-const storybookDir = path.resolve( __dirname, '../../.storybook' );
+const storybookDir = path.resolve( __dirname, '../../storybook' );
 const storyFiles = flatten(
 	storybookConfig.stories
 		.map( ( storiesPattern ) =>
