+++ tests/js/package.json
@@ -1,0 +1,21 @@
+{
+	"dependencies": {
+		"@testing-library/jest-dom": "^5.17.0",
+		"@testing-library/react": "^10.4.9",
+		"@testing-library/react-hooks": "^3.7.0",
+		"@wordpress/jest-preset-default": "^7.1.3",
+		"element-internals-polyfill": "^1.3.11",
+		"faker": "^5.5.3",
+		"fetch-mock": "^9.11.0",
+		"fetch-mock-jest": "^1.5.1",
+		"intl": "^1.2.5",
+		"intl-locales-supported": "^1.8.11",
+		"invariant": "^2.2.4",
+		"jest": "^26.6.3",
+		"jest-localstorage-mock": "^2.4.26",
+		"jsdom": "^16.7.0",
+		"jsdom-testing-mocks": "^1.13.1",
+		"lodash": "^4.17.21",
+		"use-memo-one": "^1.1.3"
+	}
+}
